+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/GreedyColoringTest.java
@@ -56,11 +56,11 @@ public class GreedyColoringTest
         Coloring<Integer> coloring = new GreedyColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(1, colors.get(2).intValue());
-        assertEquals(2, colors.get(3).intValue());
-        assertEquals(1, colors.get(4).intValue());
-        assertEquals(1, colors.get(5).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(1, colors.get(2));
+        assertEquals(2, colors.get(3));
+        assertEquals(1, colors.get(4));
+        assertEquals(1, colors.get(5));
     }
 
 }
